requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
gitpython>=3.1.40
PyGithub>=2.1.1
